feat: remove SignupListPage and update navigation to streamline customer signup management
